WIP: Invoicing, need to validate data before entering my api

+++ src/main/java/trisquel/service/ClientService.java
@@ -2,24 +2,29 @@ package trisquel.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import trisquel.Validators.Client.*;
+import trisquel.Validators.Validator;
 import trisquel.model.Client;
+import trisquel.model.DailyBookItem;
 import trisquel.model.Dto.ClientDTO;
 import trisquel.repository.ClientRepository;
+import trisquel.repository.DailyBookItemRepository;
 import trisquel.utils.ValidationErrorItem;
 import trisquel.utils.ValidationException;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 public class ClientService {
     @Autowired
-    ClientService(ClientRepository repository) {
+    ClientService(ClientRepository repository, DailyBookItemRepository dailyBookItemRepository) {
         this.repository = repository;
+        this.dailyBookItemRepository = dailyBookItemRepository;
     }
 
     private final ClientRepository repository;
+    private final DailyBookItemRepository dailyBookItemRepository;
 
     public List<ClientDTO> findAll() {
         List<Client> clients = repository.findAll();
@@ -37,6 +42,7 @@ public class ClientService {
 
     private void validateClient(Client client) {
         List<ValidationErrorItem> validationErrors = new ArrayList<>();
+
         if (client.getId() == null || client.getId() == 0) {
             client.setId(null);
         } else {
@@ -48,28 +54,27 @@ public class ClientService {
                 throw validationException;
             }
         }
-        if (client.getName().isBlank()) {
-            validationErrors.add(new ValidationErrorItem("Error", "El campo nombre es obligatorio"));
+
+        // Lista de validadores modulares
+        List<Validator<Client>> validators = Arrays.asList(new BasicFieldsValidator(), new DocumentNumberValidator(), new CuitCuilValidator(), new IvaConditionValidator(), new EmailValidator());
+
+        // Ejecutar todas las validaciones
+        for (Validator<Client> validator : validators) {
+            validator.validate(client, validationErrors);
+            // Verificar y lanzar excepción si hay errores
+            ValidationException.verifyAndMaybeThrowValidationException(validationErrors);
         }
-        //        if (client.getAddress() == null) {
-        //            validationErrors.add(new ValidationErrorItem("Error", "El campo dirección es obligatorio"));
-        //        }
-        //        if (client.getPhoneNumber() == null) {
-        //            validationErrors.add(new ValidationErrorItem("Error", "El campo número de teléfono es obligatorio"));
-        //        }
-        if (client.getDocType() == null) {
-            validationErrors.add(new ValidationErrorItem("Error", "El campo tipo de documento es obligatorio"));
-        }
-        if (client.getDocNumber() == null) {
-            validationErrors.add(new ValidationErrorItem("Error", "El campo número de documento es obligatorio"));
-        }
-        //        if(client.getEmail() == null) {
-        //            validationErrors.add(new ValidationErrorItem("Error", "El campo email es obligatorio"));
-        //        }
-        ValidationException.verifyAndMaybeThrowValidationException(validationErrors);
+
     }
 
     public void delete(Long id) {
+        List<DailyBookItem> itemsReferred = dailyBookItemRepository.findByClient(id);
+        if (!itemsReferred.isEmpty()) {
+            Set<Long> idsReferred = itemsReferred.stream().map(dbi -> dbi.getDailyBook().getId()).collect(Collectors.toSet());
+            ValidationException validationException = new ValidationException();
+            validationException.addValidationError("Error", "El cliente está siendo referenciado en los libros diarios: " + idsReferred);
+            throw validationException;
+        }
         repository.deleteById(id);
     }
 }
